+++ lib/shoulda/matchers/active_model/validation_message_finder.rb
@@ -67,4 +67,3 @@ module Shoulda
     end
   end
 end
-